changes threadproc signature to be more compatible

+++ src/cu_std/linux/tthread.cpp
@@ -42,7 +42,7 @@ void TSetThreadAffinity(ThreadID threadid,u32 cpu_mask){
     
 }
 
-TThreadContext  TCreateThread(s32(*call_fptr)(void*),u32 stack_size,void* args){
+TThreadContext  TCreateThread(s64(*call_fptr)(void*),u32 stack_size,void* args){
     
     TThreadContext context;
     pthread_attr_t attr;
@@ -57,7 +57,8 @@ TThreadContext  TCreateThread(s32(*call_fptr)(void*),u32 stack_size,void* args){
     
     _kill("failed to set stack size\n",err != 0);
     
-    err = pthread_create(&context.handle,&attr,(void *(*)(void *))call_fptr,args);
+    err = pthread_create(&context.handle,&attr,(void *(*)(void *))
+                         call_fptr,args);
     
     pthread_attr_destroy(&attr);
